Split out puppeteer-only Jest helpers due to missing `page` global
When tests were run with the Node.js or jsdom test environment, we were including helper code that tried to access the Puppeteer `page` object
