+++ lib/.eslintrc.js
@@ -1,7 +1,7 @@
 module.exports = {
   overrides: [
     {
-      files: ['jest-*'],
+      files: ['puppeteer-*'],
       globals: {
         page: true,
         browser: true
