+++ src/govuk/components/button/button.test.js
@@ -1,7 +1,9 @@
 /**
  * @jest-environment puppeteer
  */
-const { renderAndInitialise, getExamples } = require('../../../../lib/jest-helpers')
+
+const { getExamples } = require('../../../../lib/jest-helpers')
+const { renderAndInitialise } = require('../../../../lib/puppeteer-helpers')
 
 const examples = getExamples('button')
