+++ src/govuk/components/notification-banner/notification-banner.test.js
@@ -2,7 +2,8 @@
  * @jest-environment puppeteer
  */
 
-const { renderAndInitialise, getExamples } = require('../../../../lib/jest-helpers')
+const { getExamples } = require('../../../../lib/jest-helpers')
+const { renderAndInitialise } = require('../../../../lib/puppeteer-helpers')
 
 const examples = getExamples('notification-banner')
